+++ lib/chatverse.dart
@@ -8,14 +8,12 @@ export 'src/services/chat_service.dart';
 export 'src/services/auth_service.dart';
 export 'src/widgets/chat_input.dart';
 export 'src/widgets/chat_list_view.dart';
+export 'src/widgets/profile_view.dart';
 export 'src/widgets/chat_view.dart';
 export 'src/widgets/new_chat_view.dart';
 export 'src/widgets/new_chat_fab.dart';
 export 'src/widgets/chat_room_tile.dart';
+export 'src/widgets/chat_image_uploader.dart';
+export 'src/widgets/user_details_widget.dart';
 export 'src/utils/chat_theme.dart';
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}